+++ packages/nimble-components/src/chip/styles.ts
@@ -4,14 +4,12 @@ import {
     bodyDisabledFontColor,
     bodyFont,
     bodyFontColor,
-    borderHoverColor,
     borderRgbPartialColor,
     borderWidth,
     controlHeight,
     controlSlimHeight,
     iconColor,
     mediumPadding,
-    smallDelay,
     smallPadding
 } from '../theme-provider/design-tokens';
 import { display } from '../utilities/style/display';
@@ -23,32 +21,19 @@ export const styles = css`
     ${display('inline-flex')}
 
     :host {
-        background-color: transparent;
         height: ${controlHeight};
         width: fit-content;
+        max-width: 300px;
         color: ${bodyFontColor};
         font: ${bodyFont};
         padding: 0 ${mediumPadding};
-        cursor: pointer;
-        white-space: nowrap;
-        outline: none;
-        border: none;
         gap: 4px;
         background-color: transparent;
         border: ${borderWidth} solid transparent;
         border-radius: 4px;
-        fill: inherit;
-        align-items: center;
         justify-content: center;
-        outline: none;
-        position: relative;
-        transition:
-            box-shadow ${smallDelay} ease-in-out,
-            border-color ${smallDelay} ease-in-out,
-            background-size ${smallDelay} ease-in-out;
-        background-size: 100% 100%;
+        align-items: center;
         background-repeat: no-repeat;
-        background-position: center;
         ${
             /*
                 Not sure why but this is needed to get buttons with icons and buttons
@@ -59,60 +44,8 @@ export const styles = css`
         vertical-align: middle;
     }
 
-    :host::before {
-        content: '';
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        pointer-events: none;
-        outline: 0px solid transparent;
-        color: transparent;
-        background-clip: border-box;
-        transition: outline ${smallDelay} ease-in-out;
-    }
-
-    [part='start'] {
-        display: contents;
-        ${iconColor.cssCustomProperty}: ${bodyFontColor};
-    }
-
-    slot[name='start']::slotted(*) {
-        flex-shrink: 0;
-    }
-
-    [part='end'] {
-        display: contents;
-        ${iconColor.cssCustomProperty}: ${bodyFontColor};
-    }
-
-    slot[name='end']::slotted(*) {
-        flex-shrink: 0;
-    }
-
-    .remove-button {
-        height: ${controlSlimHeight};
-    }
-
-    :host(:hover) {
-        border-color: ${borderHoverColor};
-        box-shadow: 0px 0px 0px ${borderWidth} ${borderHoverColor} inset;
-    }
-
-    :host(${focusVisible}) {
-        border-color: ${borderHoverColor};
-        box-shadow: 0px 0px 0px ${borderWidth} ${borderHoverColor} inset;
-    }
-
-    :host(${focusVisible})::before {
-        outline: ${borderWidth} solid ${borderHoverColor};
-        outline-offset: -3px;
-    }
-
     :host([disabled]) {
         cursor: default;
-    }
-
-    :host([disabled]) {
         color: ${bodyDisabledFontColor};
         box-shadow: none;
         background-image: none;
@@ -130,6 +63,16 @@ export const styles = css`
         ${iconColor.cssCustomProperty}: ${bodyFontColor};
     }
 
+    slot[name='start']::slotted(*) {
+        flex-shrink: 0;
+    }
+
+    [part='start'],
+    [part='end'] {
+        display: contents;
+        ${iconColor.cssCustomProperty}: ${bodyFontColor};
+    }
+
     .content {
         text-overflow: ellipsis;
         overflow: hidden;
@@ -137,6 +80,7 @@ export const styles = css`
     }
 
     .remove-button {
+        height: ${controlSlimHeight};
         margin-right: calc(-1 * ${smallPadding});
     }
 `.withBehaviors(
@@ -161,10 +105,6 @@ export const styles = css`
                     rgba(${borderRgbPartialColor}, 0.1)
                 );
                 border-color: rgba(${borderRgbPartialColor}, 0.1);
-            }
-
-            :host(:hover) {
-                background-size: calc(100% - 4px) calc(100% - 4px);
             }
 
             :host(${focusVisible}) {