+++ packages/nimble-components/src/chip/tests/chip.spec.ts
@@ -3,6 +3,7 @@ import { Chip, chipTag } from '..';
 import { fixture, type Fixture } from '../../utilities/tests/fixture';
 import { ChipPageObject } from '../testing/chip.pageobject';
 import { sendKeyDownEvent } from '../../utilities/testing/component';
+import { waitForUpdatesAsync } from '../../testing/async-helpers';
 
 async function setup(): Promise<Fixture<Chip>> {
     return await fixture<Chip>(html`<${chipTag}></${chipTag}>`);
@@ -26,7 +27,9 @@ describe('Chip', () => {
         expect(document.createElement(chipTag)).toBeInstanceOf(Chip);
     });
 
-    it('click remove button raises remove event', () => {
+    it('click remove button raises remove event', async () => {
+        element.removable = true;
+        await waitForUpdatesAsync();
         const removeEvent = jasmine.createSpy();
         element.addEventListener('remove', removeEvent);
         const pageObject = new ChipPageObject(element);