+++ packages/storybook/src/spright/chat/input/chat-input.stories.ts
@@ -15,7 +15,6 @@ interface ChatInputArgs {
     sendButtonLabel: string;
     value: string;
     send: undefined;
-    resetInput: undefined;
 }
 
 const metadata: Meta<ChatInputArgs> = {
@@ -63,12 +62,6 @@ export const chatInput: StoryObj<ChatInputArgs> = {
             description:
                 'Emitted when the user clicks the button or presses Enter with text present. Includes `ChatInputSendEventDetail` which is an object with a `text` field containing the input.',
             table: { category: apiCategory.events }
-        },
-        resetInput: {
-            name: 'resetInput()',
-            description:
-                'Clear the input and gives it focus. This should typically be called from the `send` event handler.',
-            table: { category: apiCategory.methods }
         }
     },
     args: {